pyproject.toml
[project]
name = "csfd-db"
version = "0.1.0"
description = "Fake CSFD.cz database API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi[all]>=0.112.2",
  "pyjwt>=2.9.0",
  "aiofiles>=24.1.0",
  "sqlalchemy>=2.0.32",
  "passlib>=1.7.4",
  "types-passlib>=1.7.7.20240819",
  "types-aiofiles>=24.1.0.20240626",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "hatchling>=0.1.0",
  "ruff",
]

[tool.ruff]
line-length = 121

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["csfd_db"]
